refactor(location-manager): split foursquare client concerns
